+++ client/src/pages/DashBoard/Moderation/ReportDisplay/FilterSheet/FilterSheet.tsx
@@ -15,7 +15,7 @@ import {
 } from '@/components/ui/sheet';
 import OptionGroup from './OptionGroup/OptionGroup';
 import { REPORT_FILTTER_OPTIONS } from '@/constants/superAdmin';
-import { IReportFilterFields, TReportAction } from '@/types/superadmin';
+import { IReportFilterFields, TReportAction } from '@/types/report';
 
 export function FilterSheet({
   setReportFilters,
