+++ client/src/types/moderator.ts
@@ -1,0 +1,14 @@
+export type TModeratorRole = 'SUPERADMIN' | 'ADMIN';
+
+export interface IMderatorDetails {
+  email: string;
+  invitationStatus: string;
+  userId: string;
+  username: string;
+}
+
+export interface IModerator {
+  email: string;
+  username: string;
+  role: TModeratorRole;
+}
