+++ client/src/types/report.ts
@@ -1,3 +1,9 @@
+import {
+  REPORT_CONTENT_TYPE_OPTIONS,
+  REPORT_COUNT_OPTIONS,
+  SORT_REPORT_OPTIONS,
+} from '@/constants/superAdmin';
+
 export type TContentType = 'COMMENT' | 'POST';
 
 export interface IReportContent {
@@ -22,4 +28,23 @@ export interface IReportPreview {
   updatedAt: string;
   __v: number;
   _id: string;
+}
+
+export type TReportAction = 'RESOLVED' | 'PENDING';
+export type TReportContentType = keyof typeof REPORT_CONTENT_TYPE_OPTIONS;
+export type TSortFilter = keyof typeof SORT_REPORT_OPTIONS;
+export type TReportCount = keyof typeof REPORT_COUNT_OPTIONS;
+
+export interface IReportFilterFields {
+  action: TReportAction;
+  sortFilters?: TSortFilter;
+  countOfReports?: TReportCount;
+  contentType?: TReportContentType;
+}
+
+export interface IReportFilterOption {
+  label: string;
+  options: Record<string, string>;
+  key: keyof IReportFilterFields;
+  component: 'radio';
 }
